Added debug mode, fixed issue where copytree would crash if tempdir already existed.

+++ gillespy2/solvers/cpp/build/__init__.py
@@ -1,46 +1,52 @@
-import tempfile, shutil
+import shutil, tempfile
 from pathlib import Path
 
 from . import template_gen
-from make import Make
+from .make import Make
 from gillespy2.core import Model
 
 class BuildEngine():
     template_definitions_name = "template_definitions.h"
 
-    def __init__(self):
+    def __init__(self, debug: bool = False):
         self.self_dir = Path(__file__).parent
         self.cpp_dir = self.self_dir.joinpath("../c_base").resolve()
-        self.temp_dir = Path(tempfile.gettempdir())
+        self.template_dir = self.cpp_dir.joinpath("template")
+        self.debug = debug
+
         self.makefile = self.cpp_dir.joinpath("Makefile")
+
+
+    def __enter__(self):
+        if self.debug:
+            self.temp_dir = Path(self.cpp_dir.joinpath("build"))
+
+        else:
+            self.temp_dir = Path(tempfile.mkdtemp())
 
         self.make = Make(self.makefile, self.temp_dir)
 
-    def __enter__(self):
-        """
-        Create a new temp directory.
-        """
-
-        self.temp_dir.mkdir()
         return self
 
-    def __exit__(self):
+    def __exit__(self, type, value, traceback):
         """
         Delete the temp directory and all other associated build artifacts.
         """
+
+        if self.debug:
+            return
 
         shutil.rmtree(self.temp_dir)
 
     def prepare(self):
         """
-        Copy everything but the template definitions header to the temp directory.
+        Copy the template directory to the temp directory.
         """
 
         # If object files haven't been compiled yet, go ahead and compile them with make.
-        if not self.make.obj_dir.is_dir():
-            self.make.prebuild()
+        self.make.prebuild()
 
-        shutil.copytree(self.cpp_dir, self.temp_dir)
+        shutil.copytree(self.template_dir, self.temp_dir, dirs_exist_ok=True)
         self.temp_dir.joinpath(self.template_definitions_name).unlink()
 
     def write_template(self, model: Model, variable=False):
@@ -63,3 +69,6 @@ class BuildEngine():
         :param solver_name: The name of the solver to build. For example, ODESolver.
         :type solver_name: str
         """
+
+        self.make.build_solver(solver_name, template_dir=self.template_dir)
+
